Gohw 4

See merge request Vanek623/TaskBot!6

+++ external/task/models/task.go
@@ -7,7 +7,7 @@ import (
 
 // Task структура для хранения задачи
 type Task struct {
-	ID          uint      `db:"id"`
+	ID          uint64    `db:"id"`
 	Title       string    `db:"title"`
 	Description string    `db:"description"`
 	Created     time.Time `db:"created"`
